update,destroy自分のみに変更

+++ app/Http/Controllers/TasksController.php
@@ -110,10 +110,12 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
         
         // 認証済みユーザー（閲覧者）の投稿として更新（リクエストされた値をもとに作成）
+        if (\Auth::id() === $task->user_id) {
         $request->user()->tasks()->update([
             'status' => $request->status,
             'content' => $request->content,
         ]);
+    }
         
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -124,8 +126,11 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
+        
+        // 認証済みユーザー（閲覧者）がその投稿の所有者である場合は投稿を削除
+    if (\Auth::id() === $task->user_id) {
         $task->delete();
+    }
 
         // トップページへリダイレクトさせる
         return redirect('/');
